requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
anthropic==0.18.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock>=3.12.0
pytest-cov>=4.1.0
pytest-timeout>=2.2.0
aiohttp==3.9.1
httpx>=0.25.0
openai>=1.0.0
numpy>=1.24.0
streamlit>=1.28.0

